fix(beps): Move storeSlackThreadTs mutation to beps.ts
Convex Node.js files can only contain actions, not mutations.
Moved the internal mutation to beps.ts where it belongs.
